editor/libeditor/html/tests/test_bug966552.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=966552
-->
<head>
  <title>Test for Bug 966552</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=966552">Mozilla Bug 966552</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  var win = window.open("data:text/html,<body onload=\"document.designMode='on'\">test</body>", "", "test-966552");
  win.addEventListener("load", function onLoad() {
    win.removeEventListener("load", onLoad);
    runTest(win);
  }, false);
});

function runTest(win) {
  SimpleTest.waitForFocus(function() {
    var doc = win.document;
    var sel = win.getSelection();
    doc.body.focus();
    sel.collapse(doc.body.firstChild, 2);
    synthesizeKey("VK_BACK_SPACE", {ctrlKey: true}, win);
    is(doc.body.textContent, "st");
    win.close();
    SimpleTest.finish();
  }, win);
}

</script>
</pre>
</body>
</html>
